requirements.txt
aws-lambda-powertools==1.29.2
requests==2.27.1
python-telegram-bot==13.14
